backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cryptography>=44.0.2",
    "dotenv>=0.9.9",
    "fastapi[standard]>=0.115.12",
    "httpx>=0.28.1",
    "psycopg2>=2.9.10",
    "python-jose>=3.4.0",
    "sqlalchemy>=2.0.40",
    "sqlmodel>=0.0.24",
    "uvicorn>=0.34.2",
]
